working on basic mess tests and sketching out error handler

+++ artifact/src/base/str/streams.c
@@ -15,7 +15,7 @@ StrVec *StrVec_FromBytes(MemCh *m, byte *b, i32 length){
 }
 
 i64 Stream_ToFd(Stream *sm, byte *b, i32 length){
-    return write(sm->fd, b, length);
+    return write(sm->fd, b, (size_t)length);
 }
 
 i64 Stream_ToStrVec(Stream *sm, byte *b, i32 length){
